Add tests
refs: MBL-14856
affects: Teacher, Student
release note: none
test plan: Add attachments to a conversation
